backend/requirements.txt
python-dotenv
requests
langchain
langchain-groq
fastapi
uvicorn
ragas
datasets
sentence-transformers
pandas
streamlit
plotly